backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pydantic>=2.8.0
PyPDF2>=3.0.1
pypdf>=4.3.1
Pillow>=10.4.0
